feat(theme-generator): 适配 Web Component 模式主题切换
