+++ packages/theme-generator/index.html
@@ -7,7 +7,7 @@
     <title>TDesign Theme Generator</title>
   </head>
   <body>
-    <div id="app"></div>
-    <script type="module" src="/src/main.js"></script>
+    <td-theme-generator />
+    <script type="module" src="/src/wc-entry.js"></script>
   </body>
 </html>
