--- packages/theme-generator/src/main.js
@@ -1,4 +1,0 @@
-import { createApp } from 'vue';
-import Generator from './Generator.vue';
-
-createApp(Generator).mount('#app');
